pyproject.toml
[project]
name = "python-learning-orchestrated"
version = "0.1.0"
description = "Beginner Python learning app scaffold (architecture-first)."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Project Maintainers" }]
dependencies = [
  "google-adk",
]

[project.scripts]
python-learning = "python_learning_orchestrated.cli:main"

[dependency-groups]
dev = [
  "mypy>=1.10.0",
  "pre-commit>=3.7.0",
  "pytest>=8.0.0",
  "ruff==0.14.11",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
mypy_path = "src"
files = ["src", "tests"]

